+++ src/main/resources/fabric.mod.json
@@ -34,9 +34,9 @@
     "fabric": "*",
     "minecraft": "1.17.x",
     "java": ">=16",
-    "cloth-config2": ">=5.0.0"
+    "cloth-config2": ">=5.0.0",
+    "malilib": ">=0.10.0-dev.22"
   },
   "suggests": {
-    "malilib": ">=0.10.0-dev.22"
   }
 }
